py-srv/bin/requirements.txt
bottle
bottle-sqlalchemy
mysql-connector-python

